add webpack to import classes
