+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Tic-Tac-Toe-JS",
+  "name": "tic-tac-toe-js",
   "version": "1.0.0",
   "description": "Tic-Tac-Toe with JS",
   "main": "index.js",
@@ -18,8 +18,13 @@
   },
   "homepage": "https://github.com/Salvador-ON/Tic-Tac-Toe-JS#readme",
   "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babelify": "^10.0.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
-    "eslint-plugin-import": "^2.20.2"
-  }
+    "eslint-plugin-import": "^2.20.2",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11"
+  },
+  "dependencies": {}
 }
